refactor : separate ExamScraping Parsing Code
