+++ src/main/java/wiseSWlife/wiseSWlife/global/session/form/SessionForm.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 @AllArgsConstructor
-@Data @Getter @Setter
+@Getter @Setter
 public class SessionForm {
 
     private String sid;
