+++ src/main/java/wiseSWlife/wiseSWlife/service/graduation/scrapingInterface/TotalAcceptanceStatusScraping.java
@@ -1,0 +1,4 @@
+package wiseSWlife.wiseSWlife.service.graduation.scrapingInterface;
+
+public interface TotalAcceptanceStatusScraping {
+}
