+++ src/test/java/wiseSWlife/wiseSWlife/service/graduation/scrapingImpl/ExamTest.java
